AB#980
- Move CrewMembers to own model
- Redefine tvseason and film to use CrewMembers rather than []CrewMember
- Run `GetJobNames` and `GetMembers` on `CrewMembers`
