+++ kibble/models/film.go
@@ -31,7 +31,7 @@ type Film struct {
 	Trailers        []Trailer
 	Bonuses         BonusContentCollection
 	Cast            []CastMember
-	Crew            []CrewMember
+	Crew            CrewMembers
 	Studio          []string
 	Overview        string
 	Tagline         string
@@ -131,25 +131,4 @@ func (films *FilmCollection) MakeTitleSlugsUnique() {
 			(*films)[group[j]].TitleSlug = fmt.Sprintf("%s-%d", (*films)[group[j]].TitleSlug, j+1)
 		}
 	}
-}
-
-// GetCrewJobs - get a list of all the unique job titles in the list of crew members
-func (film Film) GetCrewJobs() StringCollection {
-	var result StringCollection
-	for _, c := range film.Crew {
-		result = utils.AppendUnique(c.Job, result)
-	}
-	return result
-}
-
-// GetCrewMembers - get a list of all the unique people who have a particular job
-func (film Film) GetCrewMembers(job string) StringCollection {
-	var result StringCollection
-	for _, c := range film.Crew {
-		if c.Job == job {
-			result = utils.AppendUnique(c.Name, result)
-		}
-
-	}
-	return result
 }
